+++ Editor/AndroidLogcatDeviceQuery.cs
@@ -1,15 +1,13 @@
-#if PLATFORM_ANDROID
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using UnityEditor.Android;
 
 namespace Unity.Android.Logcat
 {
     internal class AndroidLogcatRetrieveDeviceIdsInput : IAndroidLogcatTaskInput
     {
-        internal ADB adb;
+        internal AndroidBridge.ADB adb;
         internal bool notifyListeners;
     }
 
@@ -97,7 +95,7 @@ namespace Unity.Android.Logcat
             if (m_SelectedDevice != null && !m_Devices.Keys.Contains(m_SelectedDevice.Id))
                 throw new Exception("Selected device is not among our listed devices");
 
-            m_Runtime.ProjectSettings.LastSelectedDeviceId = m_SelectedDevice != null ? m_SelectedDevice.Id : "";
+            m_Runtime.UserSettings.LastSelectedDeviceId = m_SelectedDevice != null ? m_SelectedDevice.Id : "";
 
             if (notifyListeners)
                 DeviceSelected?.Invoke(m_SelectedDevice);
@@ -246,4 +244,3 @@ namespace Unity.Android.Logcat
         }
     }
 }
-#endif